bluetooth: Avoid leaking a fd and an infinite loop
On DLX, this code quite reproducibly receives a connection on ch_id
0 (the a2dp control channel) while it is still connected.  In
addition to the obvious fd leak caused by not closing it, this
also causes Bluetooth to go in an infinite loop.

The infinite loop comes because:

  * the leaked fd is still in the server's accept_set (select)
  * the client on this fd will close their side
  * select will keep returning that there is data on the fd but
    since we don't know about it, we ignore it and then
    immediately call select again.

Fix this by realizing that the server side channel has not yet been
closed and close it (taking care to properly keep track of the
evt flags so that we don't immediately close the new connection!).

Change-Id: Ifeaa4b3df96c6f1c0448de621e8647758676feb2